refactor: split top 2 monolithic scripts into focused modules
Splits the two oversized addon scripts flagged by the 2026-04-10
assessment into cohesive submodules while keeping public APIs
unchanged:

- optimal_control.py (392 LOC) -> facade (226) + _builders (230)
  + _config (21). Cost weights, contact/cost builders and shooting
  assembly now live alongside the public orchestration layer.
- ik_solver.py (388 LOC) -> facade (256) + _tasks (127) + _phases
  (63). Pink task/step helpers and exercise keyframe tables are
  isolated from the public solve_pose/compute_exercise_keyframes
  orchestration.

Private helpers are re-imported into the facades so existing
test mocks (patch.object, sys.modules reload) continue to work,
and no caller needs to update imports. All 478 unit/integration
tests pass.

Fixes #129
